refactor(#32): global 구조 변경
