+++ src/main/java/deepdivers/community/domain/member/service/AccountService.java
@@ -5,7 +5,7 @@ import deepdivers.community.domain.member.dto.request.AuthenticateEmailRequest;
 import deepdivers.community.domain.member.dto.request.VerifyEmailRequest;
 import deepdivers.community.domain.member.dto.response.statustype.AccountStatusType;
 import deepdivers.community.domain.member.dto.response.statustype.MemberStatusType;
-import deepdivers.community.global.mail.MailHelper;
+import deepdivers.community.global.utility.mail.MailHelper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
