+++ src/main/java/deepdivers/community/global/config/EncryptorConfig.java
@@ -1,10 +1,10 @@
 package deepdivers.community.global.config;
 
-import deepdivers.community.utility.encryptor.Encryptor;
-import deepdivers.community.utility.encryptor.EncryptorBean;
-import deepdivers.community.utility.encryptor.EncryptorTypes;
-import deepdivers.community.utility.encryptor.IpEncryptor;
-import deepdivers.community.utility.encryptor.PasswordEncryptor;
+import deepdivers.community.global.utility.encryptor.Encryptor;
+import deepdivers.community.global.utility.encryptor.EncryptorBean;
+import deepdivers.community.global.utility.encryptor.EncryptorTypes;
+import deepdivers.community.global.utility.encryptor.IpEncryptor;
+import deepdivers.community.global.utility.encryptor.PasswordEncryptor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
