+++ src/main/java/deepdivers/community/global/utility/encryptor/EncryptorBean.java
@@ -1,11 +1,10 @@
-package deepdivers.community.utility.encryptor;
+package deepdivers.community.global.utility.encryptor;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
 
 @Qualifier
 @Target({ElementType.METHOD, ElementType.FIELD})
