+++ src/main/java/deepdivers/community/global/utility/encryptor/EncryptorTypes.java
@@ -1,0 +1,5 @@
+package deepdivers.community.global.utility.encryptor;
+
+public enum EncryptorTypes {
+    PASSWORD, IP
+}
