+++ src/main/java/deepdivers/community/global/utility/mail/MailException.java
@@ -1,4 +1,4 @@
-package deepdivers.community.global.mail;
+package deepdivers.community.global.utility.mail;
 
 import deepdivers.community.global.exception.model.ExceptionType;
 import lombok.AllArgsConstructor;
@@ -10,7 +10,6 @@ public enum MailException implements ExceptionType {
 
     INVALID_VERIFY_CODE(9200, "유효하지 않은 인증코드입니다."),
     NOT_SENT_VERIFY_CODE(9201, "인증 코드가 전송되지 않은 이메일입니다.");
-
 
     private final int code;
     private final String message;
