+++ src/main/java/deepdivers/community/global/utility/uploader/S3Uploader.java
@@ -1,4 +1,4 @@
-package deepdivers.community.utility.uploader;
+package deepdivers.community.global.utility.uploader;
 
 import deepdivers.community.global.exception.model.BadRequestException;
 import deepdivers.community.global.exception.model.NotFoundException;
