--- src/main/java/deepdivers/community/utility/encryptor/EncryptorTypes.java
@@ -1,5 +1,0 @@
-package deepdivers.community.utility.encryptor;
-
-public enum EncryptorTypes {
-    PASSWORD, IP
-}
