+++ src/test/java/deepdivers/community/BackendApplicationTests.java
@@ -1,8 +1,7 @@
 package deepdivers.community;
 
-import deepdivers.community.utility.encryptor.PasswordEncryptor;
+import deepdivers.community.global.utility.encryptor.PasswordEncryptor;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
 class BackendApplicationTests {
