+++ src/test/java/deepdivers/community/domain/member/service/AccountServiceTest.java
@@ -9,7 +9,7 @@ import deepdivers.community.domain.member.dto.request.AuthenticateEmailRequest;
 import deepdivers.community.domain.member.dto.request.VerifyEmailRequest;
 import deepdivers.community.domain.member.dto.response.statustype.AccountStatusType;
 import deepdivers.community.global.exception.model.BadRequestException;
-import deepdivers.community.global.mail.MailException;
+import deepdivers.community.global.utility.mail.MailException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
